Added CollectionDocument::modify.
This provides a similar goal as the entry() API, but on documents that
were fetched at some point in the past. This was originally written as
part of an attempt to tackle #112, before realizing a fundamental issue.

This function is useful, however, so I finished adding a unit test and
decided to ship it. If you are storing a document that you've retrieved,
and you wish to perform some edits and update it, you can use this API
to edit the document in such a way that if the document has been updated
before your edit, the document will be fetched and your callback will be
re-invoked with the current document.