Implement a new backup/restore method
This implements a new version of backup/restore that will work correctly
with at-rest encryption. It still needs more work before merging.

Along the way, however, a new useful feature was developed: inserting a
document with a known ID. This is something I had avoided, because I
wanted IDs to be something that the cluster could generate in a
distributed fashion. As long as users aren't manually creating their own
auto-increment, however, there's little risk to adding this
functionality.