Added loop to wait for certificate in serve cli
If acme is enabled, we check to see if there is a certificate chain
installed before listening for QUIC connections. If running locally, a
self-signed certificate should already be installed, in which case the
certificate chain will be initialized already.

This fixes the apparently failure to listen for QUIC traffic on a
freshly deployed server when it is going from no certificate ->
certificate through acme.

Closes #115