Refactored shutdown to have a helper
This works towards #33 and #114. This new Shutdown type can be what we
expose to the Backend and plugins eventually, and it gives enough
information to each watcher that graceful shutdown can be handled across
all connection types.

That being said, this doesn't actually change any implementations -- it
refactors shutdown() to not have connection-specific code within it, and
moves that logic to the QUIC listener method.