Upgraded TransactionLog for KeyValue
This maintains backwards compatibility with the old transaction log
format, but moving forward entries will be stored using Pot, which makes
it more flexible when deserializing since it's a self-describing,
versioned format.

The rationale for separating Changes into two types of Vecs rather than
interleaving the changes in a single Vec is simply due to how this will
work after #120 is finished -- the KeyValue store will occasionally
commit a transaction with changed keys, and that process will always be
independent of the document transactions. Any attempt at tying the two
together will limit the KeyValue store's performance.