Added protocol negotation for WebSockets
Closes #47

Technically the server might be better off accepting the websocket
connection and doing a graceful termination -- as I was testing WASM I
discovered that to avoid potential abuse issues, the HTTP status code
from the failure to initialize a websocket connection isn't relayed.

So, on WASM, the websocket connection just will show up in the network
tab with a 406 Not Acceptable error.