Refactored KeyValue Expiration
In working on #124, I noticed that my original implementation of the
expiration "thread" was actually spawning a long-running blocking task.
This isn't ideal beacuse it will tie up a thread pool worker
indefinitely whenever there was a pending key expiration.

This new approach keeps the task running indefinitely, which avoids the
need for a locking mechanism around the sender. The expiration task will
end up sleeping waiting for the expiration channel to receive an update.

Finally, when the tree actually gets modified, we use a short-lived
blocking task to remove the keys.