Refactored Configuration
CHANGELOG documents the overview of the changes.

The primary motivation is making it so that adding a new configuration
option isn't a breaking change. Currently it is, because the structs
weren't non_exhaustive.

After marking them non_exhaustive, the only way to keep the API fluid is
to adopt a builder convention.

While this touches the CLI interface, I believe more changes are in
order for how to deal with configuration through the CLI.