Added ability to pre-register schemas
This paves the way for #123.

All examples have been updated to use this "schema as configuration"
methodology. Dynamic schema registration was a goal from the start for a
long-term vision of dynamically loadable plugins that define schemas.
However, it yielded this annoying "extra step" after opening the
storage/server before you could actually create a database.

This new API makes the schemas feel like part of the configuration, and
avoids the "extra step" after opening the database. Technically the step
is still there as part of the configuration building... but this feels
more correct to me now.