Making test retry on connection failure
When running under code coverage on CI, this test keeps getting a
timeout error. It works great locally, so my assumption is the listener
isn't starting quickly enough.

This loop is probably overkill -- but I'm tired of debugging CI. If this
isn't it, I'm not sure what it could be -- other examples use this port
as well, and seem to work just fine.