KeyValue operations are fully transactional now
It turns out that KeyValue operations were already transactional, but
the transactions were being committed with an empty "data" portion in
the LogEntry.

This commit now writes Changes::Keys for each operation. This is
temporary, as the goal for #120 is to make this happen periodically.
However, since this turned out to already be mostly transactional as-is,
I decided to finish up the work to getting operations in the log, and
then a separate PR can address the rest of #120.

The other note from this commit is that the internal formatting for
converting a namespace and a key to a single key for Nebari now uses a
null byte to delimit the namespace and key. Until this point, I thought
that was a good approach, but as I looked at the problem of needing to
separate the namespace from the key reliably, I realized that there was
no validation on KeyValue keys (issue coming), and because of that there
was no way to know that someone wasn't using a period in their key.

I decided that I would actually like people to be able to use a wide
variety of keys. So, I've decided to make a pre-alpha breaking change
with this internal formatting change.