Fixed expiration task circular reference
I knew when I did this originally it wasn't right, but I was in a mode
of trying to get thing working, and forgot to address this last night.

Context spawns the expiration task, and the expiration task dies when
the channel disconnects. The expiration task receiving the context means
it would never die.