Refactored KeyValue to use a manager task
This refactors the logic around persisting keys to disk by first
modifying an in-memory cache, and then persisting the changes after all
the modifications are done.

This is less efficient than before (slightly), but the next step is to allow the
manager to not always save the keys to disk, which will allow repeated
operations to keys be faster, as they won't be peristed with every
operation.