Reverted caching of open databases
This was a result of attempting to optimize an incorrect interpretation
of a profiling session. In the end, it wasn't actually an optimization
-- the extra complexity in opening a database was worse than the
overhead of constructing a new Database structure. The actual context is
already cached, and the problem with caching the Database is that it
introduces a circular reference, preventing files from ever being freed
up.