KeyValue store persisting no longer blocks
This commit's primary goal is allowing persistence to not block reads
and writes. This is done by moving dirty keys into a temporary buffer
when persisting that is shared between the persistence task and the
running state. The persistence task executes, then re-locks the state to
clear that layer of keys, which in turn unblocks future persistence
attempts.

This requires the persistence task to also check for whether the key
value state is being shut down. If so, the persistence check immediately
calls to persist again, which only does work if any dirty keys exist.

The fixes also done as part of making this work correctly:

- ExpirationLoader is now a keyed job, preventing multiple from being
  queued up if the first one hadn't finished before another Database
  instance was created.
- Key value entries now store the timestamp they were written at. This
  was already something I planned on adding for resolving conflicts when
  replicating eventually.
- ExpirationLoader now takes a timestamp of when the database is
  opening. If a key has an expiration, it now checks that the key wasn't
  modified since opening the database before updating the expiration.
- Backup/ExpirationLoader now use the same logic for iterating keys,
  which now honors dirty and pending-persistence keys. This ensures that
  backups always contain all key-value data, not just data that has been
  persisted when calling backup.