Fixed background kv worker + added concurrent test
This fixes potential issues with the background worker entering a tight
loop while persisting is happening and another key is written to. When
persisting is happening, the background worker now sleeps until it has a
new target time. The persisting operation now also updates the target
time once persisting is complete.

I added a concurrency test as well. I'm not sure if this will fix the
issues that were happening on CI or not.