Removing an extra spawn when shutting down
There was a random failure in CI that can only be explained by the
shutdown process of the KeyValue store not persisting when the tokio
runtime was shutting down in normal operation.

This change removes a spawn() call by recursing if there are any dirty
keys found after a persist operation and the shutdown flag is set.