Manually implementing histogram plot
The built in histogram plot was doing weird things when combining
metrics -- but when I used it without hdrhistogram in front of it, it
just produced weird looking charts for how much data I was providing it.

This just manually draws a bar graph, and I've checked the results and
they look a lot better.