Handling UnimpelemntedReduce correctly
Closes #25.

I originally set out to address the linked issue, but in the process I
realized that the documentation made a statement that
UnimpelemntedReduce would be returned and I caught that it wouldn't
actually be returned.

I've added a test to verify that it actually works, and in the process
uncovered some subtle error conversion bugs.