Renamed Doc and Document, more serialization work
Added CowBytes from arc-bytes in a couple of situations. The funny thing
of the state of this refactor is that we're barely using ArcBytes for
its real purpose. We still are in a few spots, but it turns out for
serialization, serde_bytes supporting a Cow-like type would have been
mostly what was needed.