Fxed multi-emit query_with_[collection_]docs
This is new functionality to the alpha, so the changelog is abbreviated
about the impact. The ability to emit multiple entries was only added
recently, and a critical simple test was missing: a query that should
match the same document twice should return both mappings.

Once this was reported (thank you @colelawrence), I realized the fix
would introduce clones of the documents. This is soemthing I found
suboptimal when using CouchDb, and I decided to introduce a new type
that would allow returning only one copy of each matched document.

The only flow that isn't clean with this implementation is if you want
to iterate over owned key/value/document combinations. Because it's
possible for multiple entries to show up, I wasn't sure how to handle
this case. Using IntoIterator is tricky because you can't borrow from
the iterator when returning from `next()`. So there's no where to hold
the BTreeMap of documents while iterating.

It would be possible to create a "unique" iterator that only returned
once per entry. But it's not that hard to do by hand, and I'm not sure
how often it will be used by anyone.