Improved doctests for collections and views
Having all the docs central seemed like a good idea at the time, but the
inline approach is much better for randomly accessing the documentation.

This moves all the existing doctests to the relevant methods, adds links
from the low level APIs to the high-level APIs, and adds a few more
missing doctests.