Removing Deref in favor of AsRef for Document
After receiving the suggestion in #177 (which I had considered at one
point as well), I decided that Deref<Header> was less useful than the
request in #177. I'm not sold on implementing Deref<C::Contents> for
CollectionDocument, but this change paves the way.

The hesitance on implementing Deref for the other type is the same
reason I was happy to remove it from Document -- it's "magic" that
sometimes breaks. When you can't move a value out because it's behind
a deref, it forces you to use the longer syntax anyways. The question
is: does deref really help that much when it sometimes can't be used
anyways?