Converting document IDs to allow any primary key
This refactoring introduces a new type, DocumentId, which allows any
`Key` implementing type to be used as a DocumentId, with the only
restriction being that DocumentIds are currently limited to 64 bytes in
length.

This restriction is to allow Header to remain copy, and allow it to not
require an external allocation.

Overall the impact to the code isn't that bad, but this is an initial
draft to have a discussion around the idea with other users.