"Fixing" a problematic test
This test very rarely fails, and the only theory I have is that it's due
to the delayed persistence for the key-value store being interrupted.
Since it requires operating a future in a runtime that's being shut
down, there's a lot that isn't guaranteed especially with task
cancellation being a thing.

Given that the persistence model of the Key-Value store is specifically
designed with this idea of data loss in mind, it's a little unfair of a
test when we don't have a reliable shutdown method to gracefully
shutdown a database before closing it.