Added Collection::PrimaryKey
This expands on the new DocumentId type to allow for a new
CollectionHeader type which contains the deserialized primary key.

There are a few issues that I have with the current setup

- Basic::document_contents(doc) is a little awkward. However, the
  previous inversion of responsibility was also weird under the hood.
  It was attractive to write doc.contents(), but it broke if
  Collection::Contents wasn't Collection. This new method always works.
- In the process of this major refactoring, I ended up making Clone
  required. I'm going to try to remove that again, or at least
  restrict it to whatever functionality needs it. It has to do with the
  Document trait and CollectionDocument implementing it -- maybe I
  shoudn't have gone that far.

Overall, I think this approach is working well enough that I'm
committing it to let more eyes review some of the ergonomics around
these concepts.