ToString/FromStr for DocumentId + Docs
All unit tests are now passing, but there's an incompatibility between
main and this branch with Views. Bincode doesn't support allowing my
custom Visitor to implement a backwards compatible path for DocumentId
inside of Header, and Views still use bincode under the hood for the
internal structures.

Ideally we would have a separate version number for the "bonsaidb"
version for views for this situation. Perhaps that's how we address it
in this branch.