Added migration path for documents
Bincode didn't support me changing the u64 type to a a custom type with
a Visitor -- what we needed to support with DocumentId. This change
introduces a new structure to parse "legacy" documents, and it's used as
a fallback.

I need to consider whether I should use this as a sign I should use Pot
to encode the Document structure. The very slight loss in speed in
theory shouldn't justify the inflexibilty -- for this exact reason.

The other major change in this commit is writing a view upgrade path
using an internal version number. After making views invalidate
themselves upon internal upgrade, I discovered a bug that affected
upgrading an existing database -- if the integrity check spawns a
mapping job, the apply_transaction operation would proceed before the
mapping job finished. There's now code to wait for any spawned mapping
jobs, which in the prositive-flow will be none.