Refactored Executed transactions to be smaller
Someone reported that they weren't able to insert as many records in a
transaction as I had expected. After some consideration, I realized
that one aspect could be optimized greatly. Each changed record had a
copy of the collection name, and this new structure keeps a list of
changed collections separate from the changed documents, and the
collection on the changed document structure is now an index into that
collection.

There is still code to be written for backwards compatibility. However,
it makes more sense to introduce a new unit test in main() that includes
a v0.1 database that can be loaded by the current version and verify its
still working.