Internalizing jobs and renaming it to tasks
This commit is in preparation of the scheduled jobs system. My original
goal was to reuse this code, but as I thought about how I was going to
implement it, it will be built atop the higher-level features of PubSub
and Collections/KeyValue to allow persistence. So this internalizes it,
and implements a simple task system for handling incoming requests in
the server separately.