Fixing task initialization on Raspberry Pi
sysinfo::physical_processor_count returns `Some(0)` instead of `None`.
This caused the number of tasks to be initialized to 0, rather than a
valid number.

This logic change takes the larger of the number of processors() or the
physical_core_count or 1, preventing potential issues on systems that sysinfo
returns 0 for both.

Refs GuillaumeGomez/sysinfo#700