Added initial importer of Open Library data dumps
This is a new example that will turn into a few things, I think.
Initially, it's a way for me to test a large amount of data. The README
shows the data sizes being tested. It served as a way for me to feel
confident in its behavior with a large, organic dataset.

To adequately do that, more than just counting the records needs to
happen. Two uses come to mind:

- Benchmarks for querying various aspects of the dataset, with other
  implementations for other database.
- Creating a command-line interface that can be used to browse/search
  the database. The way this is written could serve as a good larger
  example.

Of note, running this multiple times causes the ratings set to be
appended to each time. This is because the dataset doesn't have a
primary key for each row, most likely due to it being anonymized.