Full removing unnecessary view scan
While this was a useful check while establishing that views were working
as-intended, with Nebari we have no way to write data without it being
transactional in nature. Thus, there's no way with a correclty
implemented algorithm for this scan to ever produce any missing entries.

This could come back as an optional diagnostic/cleanup process.