Added more parallelization to view mapping
This branch started as a refactoring of the view mapping algorithm into
multiple methods rather than one huge method.

In the process, I realized that the omitted entries index was completely
unused. So, I removed that tree.

While refactoring this, I began running into borrowing issues in Nebari.
One parallelization that I wanted to do that I couldn't was allowing the
documents to be read in one thread, processed in parallel by a worker
pool, then written to the trees in a final thread. Nebari has been
updated to allow this.