Made parallelization configurable
This also changes the hardcoded 16 in the view mapper to be the number
of cpu cores by default. However, it can now be tweaked.

I also took the opportunity to update Nebari's ThreadPool to allow
setting the maximum number of threads. This configured parallelism is
passed onto the thread pool now as well.