Refactored permissions into local
Storage/Database now do permissions checks. Refs #68

StorageInstance implements StorageConnection with no permissions checks,
which is used internally to give "superuser" permissions to the
administrative APIs.

This also changes authenticate() to return a new StorageConnection. This
lays the groundwork for a single connection allowing simultaneous
authentication sessions -- something I was alraedy thinking was
necessary to support temporary role assumption (Refs #151).

As per the last commit in this branch, this work is solely focused on
bonsaidb-local at the moment. Server and Client are both in a state of
disarray.