Dispatching is now sync
This converts CustomApi to be sync only, which is a breaking change I'm
not sure I love. The backend's CustomApi might always need to be sync
for simplicity. However, I'm hopeful I can design a plugin system for
additional APIs to be defined that are able to be either sync or async,
which can be handled before a spawn_blocking call that negates the
benefits of async code.