Multiple CustomApis can now be defined
In the neverending saga of this branch, I ran into ugly design issues
with how Server was going to need to interact with CustomApi now that it
had been moved to Local.

It's been a long-term goal of mine to offer plugins that could also
define their own APIs. I realized that I could undo the addition of a
generic type to Storage/Database if I went ahead and did this work now.

That feature is still a work in progress -- it needs namespacing on the
CustomApi name, and I may refactor how ViewName/SchemaName/
CollectionName work. However, my goal is to get this branch fully passing
unit tests again before I spend time polishing everything up.