Unify internal API and custom API handling
This started as a simple project to remove actionable from the internal
networking code. It would allow for a slightly simpler design by not
requiring the Dispatcher implementation, allowing custom api handlers to
be defined as callbacks or simple functions.

In the process, I realized that it was going to be a lot of work no
matter how I approached it. I had an idea yesterday to try to switch all
API handling to the new custom API handling. One benefit of doing it
this way is that the input/output types of each internal API is now
defined in the Rust type system as associated types. This ended up
removing a ton of boilerplate code.

I still have a bit of a list before this should be considered complete.
The example is currently completely nonsensical and fails to run as a
reminder of where to pick up tomorrow.