Client's sync implementation no longer needs tokio
When initializing a client from a context where no tokio runtime is
available, and no tokio runtime handle was provided to the builder, a
new runtime is created and driven in its own thread.

In the future, if the particular transports have non-tokio
implementations, they can avoid initializing tokio entirely.

Refs #230