Changed conversion helpers for Async/Blocking
I didn't like having a From implementation that could panic, as the
documentation of such a panic invariably is hidden. There's a reaason
that From shouldn't fail.

I decided try_from seemed too verbose, so I opted to create a series of
helper functions for converting between the two types.