Limiting commerce bench in tests
Occasionally the `--all-features` test times out, and I finally found a
similarity between the times that time out and the ones that don't:
github actions is running the test with more CPUs available to it. This
scales up the test, as num_cpus is what normally would drive the number
of agents.

I don't know *why* that is too much for the CI server to handle. I don't
know why it's timing out yet. But I can't get it to fail locally despite
running it a lot on two separate machines this morning.

This change limits how many agents and shoppers there are when running
in test mode.