Wrapping some structs in Arcs
I needed to wrap the Session on Client in an Arc. In the process, I
discovered something alarming: RemoteDatabase was *huge*. A lint showing
variant size differences alerted me that a RemoteDatabase was 240 bytes.

This is alarming because I wanted all Connection/StorageConnection
objects to be cheap to move around and clone. To me, that's a few
pointers, not the equivalent of 30 pointers.

I want to address this further, but some quick investigations made me
realize I should *always* wrap a Session in an Arc to keep the impact to
the owning structures down.

The other surprise was tokio::runtime::Handle. I'm not sure how big it
is, but there are 6 struct fields, and it was some of the other bloat.
While tokio's Handle does indeed have internal reference counting,
wrapping it in an Arc allows the containing structures to remain
smaller.

I decided to commit this as its own entry so that if someone asks, "Why
is someone wrapping tokio's Handle in an Arc? It's already reference
counted!" they might discover this via git blame.

Here are the before/after numbers as reported by the new "test":

| bonsaidb_client::Client           |  40 bytes | 16 bytes |
| bonsaidb_client::RemoteDatabase   |  56 bytes | 32 bytes |
| bonsaidb_client::RemoteSubscriber | 136 bytes | 56 bytes |
| bonsaidb_local::Storage           |  48 bytes | 24 bytes |
| bonsaidb_local::Database          |  56 bytes | 32 bytes |
| bonsaidb_local::Subscriber        |  88 bytes | 64 bytes |
| bonsaidb_local::AsyncStorage      | 112 bytes | 32 bytes |
| bonsaidb_local::AsyncDatabase     | 120 bytes | 40 bytes |
| bonsaidb_server::CustomServer     | 120 bytes | 40 bytes |
| bonsaidb_server::ServerDatabase   | 240 bytes | 80 bytes |

Significant savings, but I know some of these can be improved by further
refactors.