Client sessions now deregister when dropped
As part of the switch to allow multiple authentication sessions, it
occurred to me that locally authenticated sessions are automatically
unregistered. To support remotely authenticated sessions,
ConnectedClient stores the session in its own structure, which caused
the authentication sessions to persist as long as the connection was
alive.

This commit makes the Client automatically unregister sessions when the
last copy to an authenticated session drops.

This sort of cleanup is hard to unit test, but should be. I'm going to
put in a ticket for that unit test, but for now I've manually tested
that this happens properly with debug statements.