Implemented various time key types
This started with me trying to improve the Key documentation in the
book. I realized I had a custom Timestamp time that wasn't using
variable integer encoding, and I realized that type could eventually be
deprecated and replaced by SystemTime/Duration implementations.

While doing that and coworking on Discord, it was suggested being able
to use less than 12 bytes in memory to represent various resolutions,
since the 12 bytes that Duration uses is completely overkill for
practical applications.

This set of types and traits is the result of that discussion. All
representations are a maximum of 8 bytes in memory, and when encoded
using variable encoding, the savings can be significant especially with
reduced resolution.