Refactored tuple impl macros
@ModProg was investigating implementing the Key derive, and we both
lamented how ugly the existing implementation was. We came up with the
idea to just recursively generate the bodies.

This also makes a single-element tuple a Key implementor. I originally
thought it was a useless idea, so I skipped it. However, I realized that
an esoteric use case is possible: the way this key encoding works allows
more tuple elements without breaking compatibility with the existing
stored values. I don't think anyone will ever benefit from this
particular side effect, but it's interesting enough to allow a
single-element tuple key.