Initial implementation of File structure
This has a lot of unfinished parts, but the basic insert/retrieve
operations are working to the point that I wanted to commit this.

There's something oddly satisfying about being able to use a
database-backed file using std::io::Seek and std::io::Read. This feature
alone has me excited. Seek is implemented in a completely non-blocking
way, so it will be safe to use in async code once AsyncRead is
implemented.