Revert "Refactored tuple impl macros"
This partially reverts commit a39aeb829501eb2edc724f366ec2b580d4b6c43f.

As mentioned in #237, I noticed a bug when trying to demonstrate how to
accomplish something using tuples as keys. Baffled, I checked out v0.4.0
and it worked perfectly there. I then remembered this refactoring.

The unit test for tuples was verifying consistent ordering, but it
wasn't verifying that field 1 came before field 2. This still isn't a
perfect test, but based on how these are implemented, it at least
verifies the basic functionality now.

The issue with the refactoring was this: to make macro_rules! not
complain about being ambiguous, I needed to "chop" from the front of the
invocation. This meant that the order of the arguments going into the
implementation were backwards. The last field became the most
significant, instead of the least significant.

I couldn't figure out a way around this due to the ambiguous parsing
errors. Oh well.

The unit test refactor, howevever, is perfectly fine, as the ordering
doesn't matter for that verification. The extra assert partially
verifies field ordering.