Refactored how paths work
Originally, I felt like I wanted to have a hierarchical system, which
would make moving files easier because it would only require setting the
parent id of a single object. The alternative is storing the path, which
makes it easier to compute full paths for any given document -- but
means that moving a folder will require updating every file within the
subfolders.

In the end, I realized how often getting the full path to an object is
needed, but also how less efficient (and error-prone) other mechanisms
are for implementing directory walking. Given that moving a directory is
probably a rare operation, I decided to swap out the parent_id with a
path field.

This code now supports paths fully. This virtual file implementation
treats directories and files as the same: both can contain associated
data, and both can contain children. While this may sound a little
weird, adding a unique concept of directories would purely add more code
to be tested without any real benefit that I can think of.

The permissions have been removed as the path changes also enable us to
use the built-in permissions system.