Reimplement tuple key encoding
Closes #240

This commit invents a new format for encoding composite keys. Here's a table
demonstrating the old and new format:

| Tuple          | V1          | New Format    |
|----------------|-------------|---------------|
| `("a", 2_u8)`  | `0197 02`   | `9700 0201`   |
| `("aa", 1_u8)` | `0297 9701` | `9797 000102` |
| `("b", 2_u8)`  | `0198 02`   | `9800 0201`   |

As you can see, the new format matches the ordering of the original problematic
tuples.

For non-variable Keys, the data is included directly with no additional bytes.
For variable keys, however, the CompositeKeyEncoder keeps a list of all lengths
of data being written, and places the lengths at the end of the key during
`CompositeKeyEncoder::finish()`. By moving the length information to the end of
the key, it becomes the lowest precedence for sorting.

To ensure a 'shorter' key sorts above a longer key, we add a null byte after
each variable field.

The only other tricky part is an attempt to keep the length encoding as short as
possible. By encoding the length using the same trick as UTF-8 (high-bit denotes
more data), we can encode individual lengths that are less than 128 bytes as
single bytes.

To break down the first example into its encoded parts:

- `97`: `'a'`
- `00`: variable encoding padding
- `02`: `2_u8`
- `01`: variable encoded length of "a" (1)

There is on edge case that remains with this format. Consider encoding `"a\0"`
in the above format. This can cause the comparison of the bytes to no longer be
delimited along field boundaries. As such, variable-encoded fields that contain
null bytes throw an error by default. Use
`CompositeKeyEncoder::allow_null_bytes_in_variable_fields` to forego this
safeguard.