Fixed view mapping bug when deleting entries
This is backported from f02e35e3221cc0151c340de4ee25a6f1df792c81.

A bug caused view mapping keys to be removed when the last view entry
for a key was removed, regardless of whether new entries were about to
be added. This bug was introduced in the speed optimizations in v0.3.0.

Additionally, manually backported the other fix which was hidden within
another breaking change.