Moving/renaming files is now supported
This commit also removes the idea of "directories" entirely. A specific
path prefix can be listed to only include files that directly belong to
that prefix, but there is no longer an attempt to allow parent() type
recursion up a tree. Without the ability to depend on unique view keys,
there's no way to allow moving or deleting trees/subtrees and preserving
those intermediate directory records.

This reduced feature set makes operations safer, with the only remaining
edge case being appending to a file as its being deleted -- that won't
be able to be prevented until #136.