Added MongoDB to commerce benchmark
This will either fail spectacularly in GitHub actions or simply be
ignored there for now. I need to investigate how to configure the
replSet mode in Github Actions or vet a 3rd party action to enable it --
or just give up and let mongodb on Github Actions not be benchmarked
with it enabled.

The notes on this implementation are mostly in the readme and/or
comments. Honestly, it was a pretty straightforward implementation until
I got to the materialized view, and even that turned out to be fairly
simple in the end.