Refactored blocking/async apis to be cleaner
The file type now contains the database, and it exposes methods based on
how it was originally loaded. This allows a lot of code reuse while
preventing the wrong types of methods to be called. It does mean,
however, that the File type contains the database info, which could be a
little annoying for some users. However, I suspect most of this usage
will be contained within a single method body.