Added file locking to try to guard access
Thanks to @ergho for a simple example using the key-value store that
provided a good reason to get this implemented sooner rather than later.
The small example was opening and closing the database in a loop, adding
some entries to the key-value store, and then looping. The persistence
task would eventually be writing to the database while the loop was
beginning to write to it from the main thread.

This lock should ensure that an error is raised when exclusive access
cannot be acquired.