Fixed shutdown issue in cli example test
With the new blocking file locks, we needed to ensure the database was
actually cleaned up. The previous code wasn't ever actually shutting the
server down -- the cli task would hold onto a reference to the Server,
keeping the file locked.

This change allows us to remotely shut the server down and have it
properly clean up.