Fixing panics in debug builds with locking memory
This doesn't happen reliably, but I noticed in CI there would be panics showing up in test runs, yet no tests actually failed. I tracked it down to the drop of the LockGuard, used to attempt to prevent other processes from being able to read the encryption keys held in memory. I realized one issue is that the ptr I was handing the `region` crate was a sub-range of the allocation of the EncryptionKey. By storing the key in a Box, we now have a pointer to our own allocation that isn't dependent on the host container.

Internally the region crate uses a debug_assert to check the error on unlocking.