Updating views when overwriting a non-existant key
This has always been a bug with unique views. A flag was missing being set to decide that a key had changed during an update/ovewrite operation. Specifically, this only affected a branch where an overwrite call is made against a document id that does not exist.

Credit to @rrichardson for identifying this issue.