Added ability to nest schemas
Right now this doesn't mean much, and is implemented purely at the macro
level. Long-term this should be something the Schema trait is aware of,
as schema-level lifecycle events should be triggered for these items.

For now, this is just a way to help users not need to jump into
non-derive land to use Files.